+++ src/main/java/com/sohwakmo/hr/domain/MeetingRoom.java
@@ -1,11 +1,11 @@
 package com.sohwakmo.hr.domain;
 
-import jakarta.persistence.*;
-import jakarta.persistence.criteria.CriteriaBuilder;
+
 import lombok.*;
 import org.hibernate.annotations.BatchSize;
 import org.springframework.format.annotation.DateTimeFormat;
 
+import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.Date;
 
@@ -27,7 +27,7 @@ public class MeetingRoom {
 //    private Employee employee;
 
     @Column(nullable = false)
-    private Integer reservationNo;
+    private Long reservationNo;
 
     @Column(nullable = false)
     private String title;  // 제목
@@ -54,7 +54,7 @@ public class MeetingRoom {
     private String reservationName; // 예약자
 
     @Column(nullable = false)
-    private Integer attendee;  // 참석자
+    private Long attendee;  // 참석자
 
     @Column(nullable = false)
     private Integer attendeeMax;  // 참석인원(최대 9명?)
@@ -64,7 +64,7 @@ public class MeetingRoom {
 
     private boolean state;   // 현재 상태  -- 써야하나 말아야 하나(true : 예약 중)
 
-    public MeetingRoom update(String title, String roomName, String roomPlace, String startTime, String endTime, Integer attendee, Integer attendeeMax, String purpose) {
+    public MeetingRoom update(String title, String roomName, String roomPlace, String startTime, String endTime, Long attendee, Integer attendeeMax, String purpose) {
         this.title = title;
         this.roomName = roomName;
         this.roomPlace = roomPlace;
